Remove distinction between Expressions and PropertyExpressions
Arguably, the PropertyExpression schema allows any Expression object to appear
in a PropertyExpression (using schema.self), but a PropertyExpression can only appear
in an Expression when it is used as a PropertyExpression.
We have no way of representing this in Kotlin/Java, so we just let PropertyExpression
be a type alias for Expression.